Removed IMarginExchangeAPI and placed all new logic inside IExchangeAPI per comments on the original pull request
