+++ ExchangeSharp/API/Exchanges/ExchangePoloniexAPI.cs
@@ -20,7 +20,7 @@ namespace ExchangeSharp
     using System.Net;
     using System.Threading.Tasks;
 
-    public sealed class ExchangePoloniexAPI : MarginExchangeAPI
+    public sealed class ExchangePoloniexAPI : ExchangeAPI
     {
         public override string BaseUrl { get; set; } = "https://poloniex.com";
         public override string BaseUrlWebSocket { get; set; } = "wss://api2.poloniex.com";
